Tweaked UrlJob to reuse webclient for better performance. Changed result window to show response times lower than 1 ms.

+++ Netling.Core/Models/UrlResult.cs
@@ -7,12 +7,12 @@ namespace Netling.Core.Models
         public long Bytes { get; private set; }
         public bool IsError { get; private set; }
 
-        public int ResponseTime { get; private set; }
+        public double ResponseTime { get; private set; }
         public DateTime StartTime { get; private set; }
         public string Url { get; private set; }
         public int ThreadId { get; private set; }
 
-        public UrlResult(int responseTime, long length, DateTime startTime, string url, int threadId) : this()
+        public UrlResult(double responseTime, long length, DateTime startTime, string url, int threadId) : this()
         {
             ResponseTime = responseTime;
             Bytes = length;
